Implement a basic APL ala Rob Pike's Ivy.
Obviously Rob's Ivy is written in Go, whereas this is in C, but the
architecture of this program is in large part based on my understanding
of Ivy from Rob's talk.

The lexer takes in an input FILE * and an output file descriptor.
The lexer processes the input file, dumping tokens to the file
descriptor as it goes, which is the same as Rob's lexer design minus the
differences between writing to a file versus a Go channel.
Note that the lexer uses the string module as essentially a dynamic
array while lexing tokens. (Tokens contain their string representation,
so if you're lexing a long string you want dynamic allocation). With
that said, I'm considering ripping out the string module and sticking
the dynamic array in the token itself.

The parser reads tokens from a file, and produces an abstract syntax
tree which can be evaluated or stringified.

The driver modules are example uses of the lexer and parser. The general
pattern is to fork(), have the child run the lexer and the parent the
parser.

Finally there is a test script in tests, to verify changes.