Refactor token handling to reuse allocations.
Added a prev argument of type token to functions which create tokens in
token.h. If this option is non-null, the token allocator will attempt to
re-use the previous allocation, thus reducing pressure on the memory
allocator. The idiomatic use, as shown in the update, is now to
initially pass a NULL value token, and then to re-use the previously
created value until finally you wish to free the token, so you get more
performance with less code. :)

The lexer and parsers were appropriately updated for the change.