Add support for vectors and clean up code.
Currently the vector parsing code has a dynamic array implementation
hidden in the Value type, but I'd like to move it to the parser, so that
I can simplify the Value type. (This will better align it with J / K's
value type, and also will improve performance in critical loops (parsing
is generally not critical relative to actual operations)).

This will change the make_vector() code to accept arrays and lengths,
and will get rid of the extend_vector stuff.