Take concurrency out of the lexer.
    As mentioned in the previous commit, I wanted to take the
concurrency out of the lexer and give the parser a more typical API. As
noted in the TODO.txt document, this not only allows us to remove the
serialization & deserialization code from src/token, which should
increase performance, but also simplifies the code in the parser, lexer,
and the drivers (which are now trivial).

    Despite the work it took to rip out the concurrency, I still feel
that starting with a concurrent design highlighted the cleanliness of
the API and the code internal to the lexer and parser, and having an
example of what's achievable motivates one to keep the code clean
during the refactoring.

    One remaining niggle on the topic of code cleanliness lies in the
lexer. The state machine design for the lexer emits tokens as a side
effect, which is easy when you have an infinite stream of bytes to write
to (via a file descriptor) but much trickier when you have space for
only a single token at a time.
    Originally I attempted a solution wherein one would setjmp at the
beginning of each state change, then setjmp within emit_token() and
longjmp back to the setjmp at the beginning of the state change to emit
a token. Once the token was emitted, one could longjmp back to
emit_token() and continue processing. This was interesting, probably
overcomplicated, and probably would have killed performance, but all of
that is irrelevant because it totally failed. Unfortunately,
longjmp()ing out of a function nukes its position on the stack so the
longjmp back to emit_token crashed. (It was undefined behavior anyway,
so this was bound to happen).

    As a result I went with a hackier solution, wherein I set a flag,
emitted, if there's a token to be emitted, and then on every state
change I test if the flag is set. If it's set then I unset it and return
the token, else I continue the loop. (As a side effect of this code
structure, the lexer returns an 'infinite' loop of EOF tokens upon EOF.
In any case, this setup requires adding another branch to the hot loop
of the lexer, which is kind of bad, and requires hitting the hot loop
more since only one token can be emitted from each lexer state. Thus I
have dumb code now like, for example, a left parenthesis that emits the
token and then calls the same lexer function again. I'd like to think of
a better solution when I get some time, or maybe I'll visit the people
on IRC to see what they think.

    A final interesting note was discovered while bug fixing. The lexer
actually returned a direct pointer to its internal cache, so the parser
was having trouble using it (the lexer would then change it on each
lex_token() call). As of now I changed this by (rather lazily) having
the lexer generate a token on each output. I'd like to add an input
parameter as a performance optimization to the lexer, such that that
pointer is used for the token. On the same topic, I may remove the
string library from the lexer, or the token cache. I don't think either
are that helpful in combination.