Switch token to a fixed size string.
    token now holds 2048 characters, making it fixed size. It exposes
that size in the form of token_size(), a zero-arity function. Therefore,
the lexer and parser store their tokens in-object, which removes those
dynamic allocations. That may not be the best thing for encapsulation,
as it restricts the object from growing in size dynamically, but that's
the only downside. (This setup could also be used as a sort of
small-object optimization, and then have make_token() or append_token()
return a pointer, so if the object gets too big it gets a separate heap
allocation, though I don't do that here. I believe that's how
std::vector used to work).