Bugfixes and plugged all memory leaks.
    I faced my fears with the scary parser code, which it turned out was
only scary because it had a (terrible) stack implementation where it
needed a queue. I replaced it with a small ringbuffer (which I can size
according to the parser's lookahead number), which is actually a queue,
and the code got simpler, clearer, and actually worked. I also added a
method to increment the reference counts for values, which means that I
don't leak those anymore. Finally, I fixed the handling of ASTNodes so
I don't leak those anymore either, though I do still create and destroy
a lot of objects during parsing / evaluation. I guess that's what
optimization / optimization passes is / are for though.

    Oh, I also fixed a bug in the lexer. Since we did the cheat to avoid
memset()ing the lexer's string, there are cases where the string doesn't
get null terminated (say the current lexeme is shorter than the previous
one). To solve that, we just set the null in the appropriate place
before we emit the token.