Remove ungetc from lexer with a circular queue.
    According to valgrind we've reduced the virtual instruction count by
about 200 million instructions. We're now down to 371.2 million from
894.0 million, so we're at less than half the original instruction
count (42% to be exact).

    According to my analysis, we're using around 678.7 virtual
instructions per lexeme for actual lexing. This is far from the crazy
numbers like 9 cycles per byte for symmetric encryption algorithms, and
it's unclear how directly it translates to real cpu cycles, but it's
within the realm of being not terrible as far as I can tell. Real time
is around 0.76s, for 100,001 lexemes.