Switch parser to parse strings instead of files
Switches the parser API to string-based instead of file-based.
Allows better interaction with a REPL (when the time comes) or javascript in the
form of webassembly.

Only a minor performance hit (on the order of 3% virtual instructions) when
reading from stdin (the slowest performance case with the new setup). I'd expect
performance to improve or stay the same with a direct file read via mmap().