Add statements to the grammar
This diff is large since I took the opportunity to refactor the ASTNode and
printer stuff out of the parser directory and into their own directories. As
such, there are some changes to the Makefile, which I think can be greatly
reduced by wild-carding the .o definitions.

The ASTNode now has a statement type, which is actually a list of colocated
statements (the idea being that eventually statements will be nested via
scoping, at which point statements would have children as well, and I'd probably
have to separate out the statement list and statement node types). Also
interestingly, the 'value' of a statement node is the value of the last
statement sibling, meaning that 'a = 2; b = 3;' evaluates to 3. I think this may
be useful later on for the a = b = 3 idiom from C, though I hope to have proper
multiassignment in the language.

I'm currently happy with the choice to use a stubbed internal allocator for
internally used objects, and malloc() for externally used objects. I may end up
switching to a custom pool-style allocator for the ASTNodes, once I start having
passes over the tree that delete ASTNodes (like constant folding).

Semicolon, identifiers, and let were added to the token type, as well as lexer
and proper in preparation for assignment statements. (Currently statements
themselves work, just not yet assignment statements, and as such they don't do
anything interesting yet).

I want to add proper UTF-8 support to the lexer. I'm almost there, I just need
to calculate the byte width of a given codepoint and store it so I can go
forward or backward a codepoint. I think the append* API may change to not take
an input and instead pull from the lexer's own representation of the string,
which should save the char register I use everywhere and make unicode awareness
easier (since not all UTF-8 codepoints fit in a char and possibly not an int
either (16 bit machines exist!)).

Also in the lexer, I currently have a horrible way of distinguishing between
operators and identifiers by string comparison. I'll have to think of a better
way to do this, but to be honest I want to follow Julia's footsteps and just
straight up make most "operators" user-level defined functions and thus
regular identifiers.

Finally, I added a failing test case for how I imagine assignment should work
once I implement it.