Add semicolon injection into the lexer
Borrowed semicolon injection from Go and, as a result, BCPL. This reduces user
burden in simple cases. I also removed some of the UTF-8 codepoint checks, and
advancing, because all of the tokens we check for are ascii anyway, so we don't
need them explicitly.

I also reworked the keywords check, to an array of set keywords which
identifiers are compared with. It's a much nicer and more extendable setup.