Update lexer to avoid copies.
Lexer now creates tokens which contain a pointer and a length, rather
than copying the string. This promises greater simplicity and reduced
instructions, at the cost of having the parser retain ownership of the
entire program source code for the entire compiliation.
The lexer's structure now occupies a single cache line by the way.

In ASTNode and parse, I have changed some char* pointers to const char*
pointers, which is just a form of documentation.

I've also refactored the symtable to now store void pointer values,
rather than Value values. This is not only more general, but serves as
preparation for a refactoring of the parser from recursive descent to a
Top Down Operator Precedence style which I want to do. My hope is that
such a refactoring can also get rid of the entire ASTNode branch of
code, which will be a great simplification and code reduction, as well
as potential performance improvement.

Finally, I've added an error type to the value type, which I hope can
propagate and solve the problems with reporting errors I've had up until
now.