Refactor for memory allocator usage
Updated the make file to make executables depend on their object files.
Updated TODO to reflect memory allocator requirements.

Moved most of the ASTNode creation functions into the parser. ASTNode.h
now simply exposes the structure and a few basic methods like Stringify
and Eval. Other functions, like make_value(), etc. are now inside the
parser, which I think will reduce the total code (and some fake-code
sharing, like calling several functions just to make a node). My hope is
for this to lead to overall clearer code, and fewer lines of code, at the
cost of exposing the ASTNode's internal structure. On a smaller note, I
added a portable version of asprintf() to the ASTNode directory, so that
asprintf() can be used in stringify.

The lexer has been switched to a model of simply holding types and
offsets within a buffer (which the parser presumably holds), which
shrinks the size of a token considerably, and removes copies (in the
case of types other than IDENTIFIER, the text of a token is never again
examined).

Symtable I rewrote to use the allocator, I'm thinking of switching it
back to using malloc() though, given that it has big allocation sizes.

Value now defines a vm structure, which contains memory allocators used
in the evaluate function etc.

mem.h now defines two different allocators:
a pool allocator that allocates objects of a fixed size, and creates a
free list when those objects are free()d for reusable allocations. These
objects are allocated out of slabs retrieved from the slab allocator.

The slab allocator allocates a memory block (slab) of a given size. It
then bump-pointer allocates within that slab. When the slab is
exhausted, it allocates another slab and repeats. free() is a no-op, and
all memory is reclaimed with the slab allocator itself is deinitialized.