Replace recursive descent with Pratt parser
Pratt parser has reached feature parity with the recursive descent
parser. I now need to add evaluation to the 'vm', so that it understands
scopes for functions. Once that's done, and functions are added to the
parser, completion should be a simple matter of adding the necessary
functions. Fingers crossed!